Extract `FeedImageDataLoader` expectations to a shared scope to remove duplication
