+++ FeedFetchApp/FeedFetchAppTests/FeedImageDataLoaderCacheDecoratorTests.swift
@@ -18,7 +18,7 @@ final class FeedImageDataLoaderCacheDecorator: FeedImageDataLoader {
     
 }
 
-final class FeedImageDataLoaderCacheDecoratorTests: XCTestCase {
+final class FeedImageDataLoaderCacheDecoratorTests: XCTestCase, FeedImageDataLoaderTestCase {
     
     func test_loadImageData_deliversImageDataOnLoaderSucess() {
         let (sut, loader) = makeSUT()
@@ -49,27 +49,6 @@ private extension FeedImageDataLoaderCacheDecoratorTests {
         
         return (sut, loader)
     }
-    
-    func expect(_ sut: FeedImageDataLoader, toCompleteLoadImageDataWith expected: FeedImageDataLoader.Result, from url: URL, when action: () -> Void, file: StaticString = #file, line: UInt = #line) {
-        let exp = expectation(description: "Waiting for load image data to complete")
         
-        let _ = sut.loadImageData(from: url) { result in
-            switch (expected, result) {
-            case let (.failure(expectedError as NSError), .failure(resultError as NSError)):
-                XCTAssertEqual(expectedError, resultError, "Expected load image data to fail with \(expectedError) but got \(resultError)", file: file, line: line)
-            case let (.success(expectedData), .success(resultData)):
-                XCTAssertEqual(expectedData, resultData, "Expected load image data to complete with \(expectedData) but got \(resultData)", file: file, line: line)
-            default:
-                XCTFail("Expected load image data to complete with \(expected) but got \(result)", file: file, line: line)
-            }
-            
-            exp.fulfill()
-        }
-        
-        action()
-        
-        wait(for: [exp], timeout: 1.0)
-    }
-    
     var anyData: Data { return "anydata".data(using: .utf8)! }
 }
