+++ FeedFetchApp/FeedFetchAppTests/Helpers/XCTestCase+ FeedImageDataLoader.swift
@@ -1,0 +1,31 @@
+//  Created by Ivan Fuertes on 24/07/20.
+//  Copyright © 2020 Ivan Fuertes. All rights reserved.
+
+import Foundation
+import XCTest
+import FeedFetcher
+
+protocol FeedImageDataLoaderTestCase: XCTestCase {}
+
+extension FeedImageDataLoaderTestCase {
+    func expect(_ sut: FeedImageDataLoader, toCompleteLoadImageDataWith expected: FeedImageDataLoader.Result, from url: URL, when action: () -> Void, file: StaticString = #file, line: UInt = #line) {
+        let exp = expectation(description: "Waiting for load image data to complete")
+        
+        let _ = sut.loadImageData(from: url) { result in
+            switch (expected, result) {
+            case let (.failure(expectedError as NSError), .failure(resultError as NSError)):
+                XCTAssertEqual(expectedError, resultError, "Expected load image data to fail with \(expectedError) but got \(resultError)", file: file, line: line)
+            case let (.success(expectedData), .success(resultData)):
+                XCTAssertEqual(expectedData, resultData, "Expected load image data to complete with \(expectedData) but got \(resultData)", file: file, line: line)
+            default:
+                XCTFail("Expected load image data to complete with \(expected) but got \(result)", file: file, line: line)
+            }
+            
+            exp.fulfill()
+        }
+        
+        action()
+        
+        wait(for: [exp], timeout: 1.0)
+    }
+}
